Copying //UE4/Dev-Networking to //UE4/Dev-Main (Source: //UE4/Dev-Networking @ 3289368) #lockdown Nick.Penwarden #rb none
==========================
MAJOR FEATURES + CHANGES
==========================

Change 3240806 on 2016/12/20 by John.Barrett

	Added 'SendUnitRPCChecked' function, to streamline executing 'UnitTestServer_' pseudo-RPC functons within unit tests, on the server.

	Worked around an Engine bug with NetConnection.QueueBits.

Change 3240807 on 2016/12/20 by John.Barrett

	Added new variations of 'ReceiveNetFieldExports' exploits, "ReceiveNetFieldExports_MapAlloc", "ReceiveNetFieldExports_NullAlloc", "ReceiveNetFieldExports_NullCheck".

Change 3241838 on 2016/12/21 by John.Barrett

	Quick compile fix for Win32.

Change 3243149 on 2016/12/22 by John.Barrett

	Another Win32 compile fix.

Change 3252022 on 2017/01/10 by John.Barrett

	Added new non-shipping commandline switch, for appending to the commandline using a text file: -CmdLineFile="D:\Users\John\Desktop\ServerCmdLine.txt"

	This sidesteps Windows commandline length limit.

Change 3257927 on 2017/01/14 by John.Barrett

	Fixed a couple of issues in ReceiveNetFieldExports, and deprioritized the remaining issues by only allowing execution when Replay is enabled.

	#JIRA UENET-460, UENET-466, UENET-472, UENET-478, UENET-484, UENET-490

Change 3258731 on 2017/01/16 by John.Pollard

	Refactor network actor list management to be more efficient

	* Move dormancy list management to FNetworkObjectList
	* Optimize actor network dormancy by removing actors from the active list that are dormant on all connections
	* Removed NetUpdateTime on actor, and now use the NextUpdateTime on FNetworkObjectInfo (these values are more hot in the cache too)
	* We now early out of the consider logic faster when possible
	* Remove other misc unused network state/code and general cleanup

	Merging using FNMain->DevNetworking

Change 3258736 on 2017/01/16 by John.Pollard

	Fix compile

Change 3259482 on 2017/01/16 by John.Pollard

	Deprecate GetNetworkActor

Change 3264957 on 2017/01/19 by John.Pollard

	Deprecate bPendingNetUpdate, NetUpdateTime and LastNetUpdateTime

	Merging using FNMain->DevNetworking

Change 3264962 on 2017/01/19 by John.Pollard

	Fix for FORT-35711 - Edited buildings do not always replicate correctly

	We were removing the actor from the network object list too soon

	Merging using FNMain->DevNetworking

Change 3264963 on 2017/01/19 by John.Pollard

	Turn on adaptive network updates by default

	Merging using FNMain->DevNetworking

Change 3265878 on 2017/01/20 by John.Pollard

	UE-40902 - CIS warnings due to deprecated properties

Change 3265930 on 2017/01/20 by John.Pollard

	UE-40902 - CIS warnings due to deprecated properties

Change 3273552 on 2017/01/26 by John.Barrett

	Added way to specify Associativity for binary operators with the expression parser, which defines whether operators of the same precedence are executed Left-To-Right, or Right-To-Left.

	Defaults to Right-To-Left for legacy reasons.

Change 3273553 on 2017/01/26 by John.Barrett

	Added 'reflect' console command, tying existing FVMReflection helper to a string parser, allowing full access to UE4 reflection using C++ style syntax (like a supercharged get/set command), e.g:
	"Reflect Find(,PlayerController).Player.ViewportClient.GameInstance.LocalPlayers"

	Added supporting FVMReflectionParser class, for evaluating strings as reflection statements, using the ExpressionParser code to implement the syntax.

	Added automation testing for ExpressionParser binary operator Associativity (placed here rather than in Core, as the reflection parser is better suited to associativity tests).

Change 3273555 on 2017/01/26 by John.Barrett

	Added 'ServerAddCardInternal' and 'ServerUndoRemoveCard' exploit unit tests, for testing more nullptr access exploits in the card RPC code.

	Updated expected results on unit tests for issues that have been fixed.

Change 3273931 on 2017/01/26 by Bart.Hawthorne

	- Steamworks has been upgraded to 1.39
	- The steam controller implementation has been rewritten to accommodate the new interface added in 1.36

Change 3275010 on 2017/01/27 by John.Barrett

	Quick compile fixes.

Change 3276218 on 2017/01/27 by John.Pollard

	FORT-36482 - Fix issue with using wrong serializer for re-mapping objects

	Merging using FNMain->DevNetworking

Change 3276219 on 2017/01/27 by John.Pollard

	First pass at implementing net relevancy for replays

	* All connections are considered when determing if an actor is relevant
	* Enable by setting demo.UseNetRelevancy to 1
	* Override cull distance with demo.CullDistanceOverride

	Merging using FNMain->DevNetworking

Change 3278725 on 2017/01/31 by John.Barrett

	Fixed multiple card system RPC exploits, through checking parameters in _Validate functions.

	#JIRA UENET-445, UENET-512 and UENET-518

Change 3278730 on 2017/01/31 by John.Barrett

	Misc. NetcodeUnitTest updates.

Change 3278733 on 2017/01/31 by John.Barrett

	Marked last of card unit tests as fixed.

Change 3278956 on 2017/01/31 by Bart.Hawthorne

	Disable static analysis warnings from Steamworks headers

	#jira UE-41262

Change 3278978 on 2017/01/31 by Bart.Hawthorne

	Fixed Linux warning caused by initializing variables out of order

	#jira UE-41307

Change 3280081 on 2017/01/31 by Bart.Hawthorne

	ShooterGame now supports dedicated servers on PC/Mac/Linux

Change 3281217 on 2017/02/01 by John.Pollard

	PR #3172 - GitHub 3172 : Fixed IPAddressBSD validation when setting Ip from string.

Change 3283556 on 2017/02/02 by John.Barrett

	CIS compile fixes.

Change 3284509 on 2017/02/02 by John.Barrett

	CIS no-PCH compile fixes.

Change 3285075 on 2017/02/03 by Bart.Hawthorne

	- Use Engine\Binaries\ThirdParty\Steamworks libraries for 64 bit Steam builds since it seems to be working with 1.39
	- Add 32 bit and 64 bit Steam libs for linux

Change 3285229 on 2017/02/03 by John.Barrett

	CIS monolithic include warning fixes.

Change 3286108 on 2017/02/03 by Bart.Hawthorne

	- Fixed issues with Steam not working correctly on Linux
	- Removed unnecessary controller.vdf dependency

[CL 3289494 by Ryan Gerleve in Main branch]

+++ Engine/Plugins/Online/OnlineSubsystemSteam/Source/Private/OnlineUserCloudInterfaceSteam.cpp
@@ -352,7 +352,7 @@ bool FOnlineUserCloudSteam::RequestUsageInfo(const FUniqueNetId& UserId)
 
 void FOnlineUserCloudSteam::DumpCloudState(const FUniqueNetId& UserId)
 {
-	int32 TotalBytes, TotalAvailable;
+	uint64 TotalBytes, TotalAvailable;
 	if (SteamRemoteStorage()->GetQuota(&TotalBytes, &TotalAvailable) == false)
 	{
 		TotalBytes = TotalAvailable = 0;
